Change preference type from float to decimal
Fixes #57

+++ app/models/spree/app_configuration_decorator.rb
@@ -1,3 +1,3 @@
 Spree::AppConfiguration.class_eval do
-  preference :use_store_credit_minimum, :float, :default => 0.0
+  preference :use_store_credit_minimum, :decimal, :default => 0.0
 end
